"4:55pm.
```
inl as_cost =
    bind <| inl x ->
        inl out = primal x
        {
        out
        bck=inl _ -> x 0 <- one
        }
```

Piece of cake. Now I can just pass the TD cost through this. No need to go through complications in the TD function.

5pm.

```
    inl td {r discount_factor eligibility_decay R' V' V} =
        inm R = r + discount_factor * (eligibility_decay * R' + (one - eligibility_decay) * V')
        {R' = R; out = sqr(R - V) |> as cost}
```

What I really like about this formulation is that there will be gradient flow through `R'`...though I do not think that will affect anything.

...I missed some things...

5:05pm.

```
    inl td {r discount_factor eligibility_decay R' V' V} =
        inm R = r + discount_factor * (eligibility_decay * R' + (one - eligibility_decay) * V')
        inm error = R - V |> as_cost
        {R error}
```

Let me do it like this. Let me not square it. The back of sqr is just `2x`. That means it would just multply the 1 feed at the top to two and that is it. What I wrote is equivalent to `sqr(R - V) / two`. Actually, let me multply it by two.

...No wait, that would also multply the reward feed to the actor. Nevermind. Let me just to it like this. I'll just increase learning_rate of the critic by 2x instead later in its layer's optimize.

Agh, how complicated. Here is a list of things I need to do for the weights in the AC layer.

- On the front side, project it based on the inputs.
- On the back side, project only the `action_probs`.
- Increase the LR for the critic by 2x.
- In addition to that worry about the biases...I think I am just going to skip having these in the final layer.

...You know, how about I just apply KFAC all together on the back side. Right now I might be tying myself into a loop in trying to figure out how to set the learning rates.

Actually, I still do not understand how the backward pass of KFAC is suppose to work. It is not just TD that I am worried about - I mean wouldn't rescaling the gradients ensure that the network never converges.

If the gradients are small, maybe they are suppose to small right?

Maybe this is just an artifact of KFAC. Maybe with probabilities it really makes sense to KFAC the back.

5:50pm. I've had some time to think about this during lunch. Yeah, with probability distributions KFACing the back would make sense. With linear layers not so much.

And maybe representing the output of the TD layer as a probability distribution would take of the scaling thing. I should look up stuff on Bayesian RL. There was that survey a while back. I will surely find something on Bayesian actor critics.

Without a doubt, KFACing the whole layer is what I should be doing.

It would be without a doubt the most stable and sensible choice of all.

6pm. Damn, I am tired. I do not think I will coding anymore today. I am finally getting back into the swing of things.

Right now things are going really, really well. I believe in the idea that I have now and the benefits it will bring so I have no more reason to hesitate. I just go out there and grasp my victory.

The same goes for plastic RNNs.

By the time the next update rolls around I will have done some excellent work."