I know I was supposed to start out today guns blazing, but some of the most trivial things are giving me huge conceptual difficulties.
Like those 3 bias terms in the plastic RNN. I am only half joking when I say that I couldn't sleep properly because of thme last night.

What I've decided to do because of that is pretty radical. I am going to make them pruned neurons. During the optimization stage I am going to apply a mask on everything specifically to zero them after the update.

For convolutional nets I've been also having trouble deciding, but I've settled on the simple choice of having them be separate weights.

From here on out, biases will be important parameters - they will be there to take the place of what would otherwise be hyperparameters.

I am not going to use them like I did in MI - each one of them will be an individual scalar.

Somehow inadvertedly, this whole experience of the past two months has really taught me the importance of pruning. I can really see how adding extra weights to the biases dependent on the inputs leads to a spike in variance. Since Pruning is good for the biases it is highly like that it would be good for other things as well.

There probably isn't a better variance reduction technique possible than pruning. There are all sorts of schemes going on now in RL, but those guys are not at all thinking about the most obvious one.

If I could find a way to automate pruning, I could reap a huge benefit.

Interestingly, I have something to begin with. Originally KFAC was made to investigate pruning so I'll be able to take advantage of it for that purpose. The EigenNet paper got a good gain from that.

...

It was not at all my intention to start doing this here - I just want to prune the bias terms {upper mid alpha} because it is much easier to arrange everything that way, but the flow of events is simply pushing me in this direction.

...

Right now I am thinking how I would apply to convolutions. In order to simply my thinking I am transforming the inputs suitable for matrix multiplies. It is rather clear cut actually.

The biggest problem will be figuring out how to do the covariance matrix updates, but I'll handle that somehow.

...

Also, I 've figured out KFAC's backward pass. I finally remembered that it converges on Mnist, meaning that even though it cannot tell apart the difference between poor init and being close to convergence, that does not matter.

This was causing me a lot of trouble.

One thing I am absolutely sure is that I want everything in a layer to be on the same scale. This is despite the fact that I did not get any benefit from applying KFAC on biases last time.

Well, to be fair, given my experience of the past 3.5 years, I'd say that I did not get much benefit from biases at all. Hence I should just constrain them to what would otherwise be important hyperparameters. I'll centralize everything as much as possible, and in convolutional nets since it will be impossible to do that I'll bring back the old `expand_singular` trick.

Probably, prunning will be the main way to defeat catastrophic forgetting.

...

One more thing before I take a break from all this thinking. Given how unlikely that LSTD-style of model-based RL for deep nets will ever be found, controllable eligibility traces might be the closest missing link between model-free and model-based RL. Plasticity will give it wings, but realizing this by no means the task is over. The task of how to reduce variance will only start when I do this. 2019 will have that as a theme. There are vast gains possible in this direction that haven't yet even begun to be tapped into.

Yesterday I read the soft AC paper - there are still gains to be had in this direction, but that sort of thing is not cutting to the meat of the issue.

Well, I'll consider the soft AC when I start with real money - off policy learning will definitely be a boon there.

I like how it does not need target nets much - with KFAC that will go all the way. I'll aim to implement it at some point in the future, but it is not a priority at the moment as I will be aiming for online learning with RNNs.

...

My goal for today will be to make the dedicated AC layer. I realized yesterday that I do not need a log cost - literally everything can be turned into a NG method with the `log << exp` trick. You just exponentiate the cost and then log it - this is an identity operation, but the math for NG should all apply in the end. I do not have to worry about this at all. I can use KFAC with squared error with no problem as one would expect.

The interesting thing about the AC layer is that its backprop step will also have a backprop step. It is interesting, first the graph goes right, then all the way back to the left in the head. Then this is reversed during the backprop stage. This is because I need to backprop into the `eligibility_decay`. In fact that should have been the correct way of doing it, but did not occur to me last time. There was no need for it then, but there is now.

https://www.frontiersin.org/articles/10.3389/fnins.2017.00324/full
https://www.sciencedirect.com/science/article/pii/S0893608097000117
https://papers.nips.cc/paper/472-competitive-anti-hebbian-learning-of-invariants.pdf

This last paper is interesting, so let me take a break here and I will do some proper work when it is time to do so.