Let me pry myself off from slacking (though I am still thinking about things) and I will start work. I need to shake my indecision and make the decisive steps forward. Yesterday towards the end I got blocked up by various issues, but now I should be able to go past them.
In the end, applying KFAC to the entire layer is the way to go. This is what I should have done from the start, but without the tree views there is no way I'd ever even try attempting it. It is true that tree views for tensors are a massive library feature.

Let me gather my will and I will start.