" ``` 1/2   0   1   0   2 1/2   1 1/2   3 ```
This leads to...

```
 46/7  4/7 -16/7
  4/7  4/7  -2/7
-16/7 -2/7   8/7
```

Actually, this is more non-linear than I'd expected it would be. I did not think those zeroes would become positive.

Hmmm...actually this was rather illuminating. I am satisfied. I can go forward with this.

Let me sketch out the layer. I'll base it on the RNN one. The feedforward layer right now is useless. It has nothing that I'd need.

1:50pm. Right now I am thinking how I could replace the `sampling_pg` with something more elegant. Even if I did so, at the bare minimum, I'd need a kind of map that would allow me to store scalars.

So I'd have to eschew a map and do my own loading and storing in `iter`, but it would be acceptable for a one of specialized function such as this one. The first step towards this would be to put the sampler into `CudaAD`'s `Seq` module. But I'll do that after I get RL operational again. There is no need to innovate now only to have difficulty debugging it later."