"2:30pm.
```
        inl ac_sampling_action {action_probs eligibility_decay V} s =
            inl {out bck} = sampling_pg action_probs s
            {
            out
            bck=inl d ->
                inl {out={R error} bck=bck'} = map CudaAD.Activation.td {d with eligibility_decay V}
                {
                out={R'=R; V'=V}
                bck=inl _ -> bck' (); bck {reward=error}
                }
            }
```

This function itself also came out simple to the limit.

2:35pm. Now what is next? The layer is essentially done thanks to this.

2:40pm. `{action ac}`. This is fine.

2:45pm. I am just taking a short breather to think about it.

No matter how I look at it, I've done everything I wanted in the `Learning` module.

It is time to move on to poker. There I will have to take account for the new backward pass.

It is time to get to it."