"2:50pm.
```
    inl player_pg {name actor learning_rate} cd =
        open Learning
        inl action = RL.action {State Action final=RL.sampling_pg}

        inl input_size = Union.length_dense State
        inl num_actions = Union.length_one_hot Action

        inl net,_ = Tuple.append (Tuple.wrap actor) (RL.Layer.pg {size=num_actions} :: ()) |> init cd input_size
        //inl net,_ = Tuple.append (Tuple.wrap actor) (Feedforward.linear num_actions :: ()) |> init cd input_size

        inl run =
            Union.mutable_function
                (inl {state={state with net} input={input cd}} ->
                    inl {action net bck} = action {net input} cd
                    inl bck x =
                        Struct.foldr (inl bck _ -> bck {x with learning_rate}) bck ()
                        Struct.foldr (inl {bck} -> Struct.foldr (inl bck _ -> bck {learning_rate}) bck) net ()
                    {state={net bck}; out=action}
                    )
                {state={net}; input={input=State; cd}}

        inl methods = {basic_methods with
            bet=inl s input -> s.data.run {input cd=s.data.cd}
            showdown=inl s reward ->
                inl l = s.data.run.reset
                inl reward = dyn (to float32 reward)
                List.foldl (inl _ -> function {bck} -> bck {reward} | _ -> ()) () l

                Optimizer.standard learning_rate s.data.cd s.data.net
            game_over=inl s -> ()
            }

        Object
            .member_add methods
            .data_add {name; win=ref 0; net run}
```

I'll start by doing some cleaning. I am not going to be using this one anymore that is for sure.

Goodbye PG based player. You served me well.

```
inl player_mc_ac {d with name learning_rate discount_factor} cd =
```

I'll rename this one to just AC. I am not going to be doing MC anymore.

I wonder if I am doing the right choice. I think for poker TD(0) might be ideal in the end...but actually in the end TD(0) would be ideal for everything. It is fine if it starts at two step and then pushes the eligibility up or down depending on the circumstances. In truth, I'd prefer it to start at 1 instead.

...Let me set the bias for eligibility trace to 2.0. MC at the start would be better. Or maybe even 3?

...Ah, I do not have a bias. Well, I guess I'll do it directly in the map function then.

3:30pm. Had to take a mighty break. Wow, I really was away for quite a while.

```
    inl player_ac {d with name learning_rate discount_factor} cd =
        open Learning
        inl action = RL.action {State Action final=RL.sampling_pg}

        inl input_size = Union.length_dense State
        inl num_actions = Union.length_one_hot Action

        inl learning_rate = {
            actor=learning_rate
            critic=learning_rate ** 0.85f32
            shared=learning_rate
            }
```

Ah, I forgot about layer sharing and such. Yeah, I won't be needing this anymore. All the gradients will be unblocked.

I am thinking about whether I should also speed up the learning of the eligibility trace. Is that part of the critic or a part of the main net?

I am leaning towards it being a part of the main net. After all I was seriously considering making a direct output of the lower layers, but have changed my mind at the last minute. So it will have a regular learning rate then, though I'll make sure to experiment.

Now since I am using KFAC, I am going to have some difficulty actually passing in the learning rate for the appropriate section. Well all the reprojections are linear so I can map with the reduced lr the critic part before passing it along.

That is the easiest choice here.

3:45pm.

```
inl ac = RL.ac num_actions |> init cd input_size |> fst
```

I've replaced a fairly crazy amount of code with just this.

```
inl action = RL.action {State Action final=RL.sampling_pg}
```

Almost forgot to remove this.

...No wait, actually I need to keep this in. But I think I will just do it directly in the layer.

...No wait, I am doing it directly in the layer already. That statement is just a thump I forgot to get rid of months ago.

3:50pm.

```
                    inl ac, out = run cd input ac
                    inl bck x =
                        inl {value bck=critic_bck} = RL.Value.mc critic_out cd x
                        critic_bck {learning_rate=learning_rate.critic}
                        actor_bck {reward=value}
                        inl f learning_rate network = Struct.foldr (inl {bck} -> Struct.foldr (inl bck _ -> bck {learning_rate=learning_rate}) bck) network ()
                        f learning_rate.critic critic; f learning_rate.actor actor; f learning_rate.shared shared

                    inl action = Union.from_one_hot Action (cd.CudaTensor.get (out 0))
                    {state={actor critic shared bck}; out=action}
```

There is actually quite a lot of stuff to remove. Well, let me do it as it should be and then I'll adjust the tests in response.

3:50pm.

```
                    inl ac, out = run cd input ac
                    inl action = Union.from_one_hot Action (cd.CudaTensor.get (out 0))
                    {state={ac}; out=action}
```

I am a bit confused here. Is bck a part of the `ac` or a part of `out`? Ah, it is definitely a part of the net. Ok.

```
        inl run =
            Union.mutable_function
                (inl {state={state with ac} input={input cd}} ->
                    assert (eq_type State input) "The input must be equal to the state type."
                    inl input =
                        inl tns = Union.to_dense input |> Tensor.array_as_tensor
                        cd.CudaTensor.from_host_tensor tns .reshape (inl x -> 1, Union.length_dense State)
                    inl ac, out = run cd input ac
                    inl action = Union.from_one_hot Action (cd.CudaTensor.get (out 0))
                    {state={ac}; out=action}
                    )
                {state={shared actor critic}; input={input=State; cd}}
```

Then this is all I need in run.

3:55pm.

```
            showdown=inl s reward ->
                inl l = s.data.run.reset
                inl reward = dyn (to float32 reward)
                inl original = discount_factor
                List.foldl (inl discount_factor x ->
                    match x with
                    | {bck} -> bck {discount_factor reward}
                    | _ -> ()
                    original * discount_factor
                    ) discount_factor l
                |> ignore

                inl cd = s.data.cd
                inl f learning_rate = Optimizer.standard learning_rate cd

                f learning_rate.shared s.data.shared
                f learning_rate.actor s.data.actor
                f learning_rate.critic s.data.critic
```

This is all going to have to be redo from scratch. In fact, I am not sure how to do the double `bck`.

```
                inl original = discount_factor
                List.foldl (inl discount_factor x ->
                    match x with
                    | {bck} -> bck {discount_factor reward}
                    | _ -> ()
                    original * discount_factor
                    ) discount_factor l
                |> ignore
```

Ah, note how I am reducing the discount factor by hand here. That is just such a stupid way of doing this. The current way of doing eligibility traces is so much better. And it would be impossible without the `CudaAD` stuff I did very recently. Investments in compiler technology pay of in spades.

4pm. You know, let me just write out how I want it to be without worrying about the types and then I will move from there.
"