"4:15pm.
```
            showdown=inl s r ->
                List.foldl' ignore (inl next {r R' V'} x ->
                    match x with
                    | {bck} ->
                        inl {out={R' V'} bck} = bck {discount_factor r R' V' }
                        next {r=dyn 0f32; R' V'}
                        bck ()
                    | _ -> ()
                    ) {r=dyn (to float32 r); V'=0f32; R'=0f32} s.data.run.reset

                Optimizer.standard learning_rate s.data.cd s.data.ac
```

Here I think I will just use the stack to simplify things for me. No need to make everything generic.

Ah, no wait, I am sort of assuming here that `bck` includes only the last.

What was that `foldl_map` thing I did last time? It was in `learning2`. I need to bring it out.

4:20pm. No I messed up. I am going to need two nets. `ac` will have to be a separate layer otherwise all of this will be too complicated.

4:30pm. Shit. I went the wrong way. I totally was not thinking about how the final layer needed special treatment.

Also I am not sure at all how am I going to speed up the critic. Previous it was not only it that I speed up. Since it was on its own layer, the separate learning rate would affect how the covariance matrix was handled too.

Well, input covariance should probably learned at the same rate the actor does. Since it is the actor that will indirectly affect that.

And since I have them on the same layer it does not longer make any sense to have a separate learning rate for the back covariance.

That leaves only the critic itself to fiddle with, but maybe for once I am better off just syncing it with the rest now that I have KFAC in the same layer. I can always speed it up later if necessary.

Let me not worry about that.

What I am going to do now is take the activation out of the AC layer. I'll leave it separate."