"9:25am. file:///E:/Downloads/Metaheuristics/RL/Natural%20Temporal%20Difference%20Learning.pdf
```
For a related
application, Amari (1998) suggests a transformation of a
parameterized function to a parameterized probability distribution.
Using this transformation, the Fisher information
matrix is
```

9:40am. file:///E:/Downloads/Metaheuristics/Higher%20Order%20Optimization/Natural%20Gradient%20Works%20Efficiently%20in%20Learning.pdf

This stuff on 6/36 might be exactly what I need.

9:45am. Ah, it is not `log <| exp x`, but `log <| exp -x`.

But it does not really matter, the really important thing is that normalizing constant that he slinked in.

9:55am. Yeah, I am not getting anywhere here. What I need is the network to predict the scale of the reward somehow. That is the only choice.

10:10am. https://www.reddit.com/r/reinforcementlearning/comments/9w2gny/d_in_td_learning_is_there_a_way_to_have_the/?st=jocnsep2&sh=74e5418c

"When applied to square error, natural gradient methods essentially have a scaling constant (see page 6/36) in front of them. Given that I want to do RL with NG, it seems sensible to imagine that the network itself should predict the scale of the reward and take advantage of that when doing updates. Surely it should be possible to do more than just predict the mean. Manual reward scaling is complete nonsense because even with NG it would act as a multiplier on the critic's learning rate.

Had any work been done towards that end?"

I've studied this a bit and will see what answers do I get. So the answer is `log exp` in the end, but that introduces the scaling constant. Damn.

Ok, forget it. I'll rescale the rewards so that they are in the [-1,1] range. There is no point in worrying about this. Let me do some programming.

10:30am. https://stats.stackexchange.com/questions/7440/kl-divergence-between-two-univariate-gaussians
https://link.springer.com/article/10.1007/BF00133173
Intrinsic losses

Agh, forget programming. I cannot program while something is weighting on my mind. Let me study this."