https://www.youtube.com/results?search_query=cross+entropy&page=&utm_source=opensearch
I'll look up stuff on cross entropy in a bit. Right now I am still thinking about the univariate Gaussian idea. It is THE way to do this. There cannot possibly be anything else. Squared error is obviously the wrong cost to use in the critic, much like having a fixed eligibility decay is.

I'll combine it with the bounded exponential activation that I made for modulating the plastic RNN.

Just like it has been repeatedly happening for the past week, right now I am crazy over this. I feel such a powerful surge of inspiration.

I did it! I've solved two of the major obstacles to doing model-free RL!

I have no doubt that the tests I will do will see me through and prove the worth of my ideas.

I am going to have to do the AC from the top in order to accommodate the new reality of how things should be done - I am going to have to sacrifice some speed for convenience. I will add the bias to the input and will implement the mask idea for `{upper mid scale scale_fin}` parameters.

Ahhh, things are going to be so wonderful when all of this is done.

With TD learning done properly, I will be able to rest more easily and leave the critic out of mind to focus on the plastic RNNs. There is no doubt that the two ideas I will make will give me a significant benefit in sample complexity just on their own. That recent paper said that PPO was not learning the true value function. Well, this thing certainly will be doing that.

Simply put, without the scale prediction, the updates will be seriously off model.

I am guessing that this has been tried before and has been found to be unstable, but it likely that they missed the bounded exponential and KFAC ideas. These two will make it work without a doubt.