"It is time to resume.
https://math.stackexchange.com/questions/1028224/is-there-a-symmetric-alternative-to-kullback-leibler-divergence

Yeah, I think that it might be a good idea to use the symmetric distance.

https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence
"In simplified terms, it is a measure of surprise."

Yeah, this is what I need. You know, maybe I'll go with the standard thing. I am comparing a sample to a model so I should use the standard one.

2:20pm. I feel like studying right now. Since the iron is hot I might learn a little about information theory and such.

Let me watch some videos."