"3:25pm. I've realized something. Once I have the scale, I can use it to divide the reward given to the actor. Honestly, this is the only realistic thing I can think of how to make RL algorithms work on tourneys. HUSNG might be good unscaled, but tourneys will have just too large of a scale.
Rescaling makes sense if you consider TD(0) - every episode is unique and so can have its own reward.

With this I would bridge the gap between model-free and model-based learning in one fell swoop.

I'll do it all step by step. I will build up all the pieces for my victory.

https://www.youtube.com/watch?v=UrefKMSEuAI&list=PLE125425EC837021F

I guess I'll spend a while watching this playlist while I let my mind digest the things I thought up today. I am in the mood for this kind of thing. I need to evaluate and reevaluate whether it makes sense to rescale the reward depending on the input.

Linearly, a huge reward would mean that you'd want to move the policy to 100% immediately for that action, but seen globally I am not sure whether that is the conclusion that I want to take. The scale of the reward differs depending on the context.

Ideally you'd want one net for each context, rather than a single net for the whole.

That is what rescaling the rewards for the actor would get me.

3:50pm. Well, maybe. I am still not entirely the sure what the expected behavior for this should be.

...Let me just watch those videos. Two more hours of this and I will take a break."