It has been over 3.5 since the start of the journey and now I finally understand cross entropy, entropy and KL divergences.
I am convinced now, this is something worth using everywhere. The univariate Gaussian KL loss can replace squared error everywhere I can think of. There is literally no reason to use it.

On the surface, natural gradient methods might have the disadvantage of needing log p cost function, but that is actually a blessing in disguise.

I just have to make it work. The tricky part is the scaling and with bounded exponential functions it could be done.

Using KFAC will clinch this - it cannot possibly fail.

Tomorrow, I will forget the work that I did up to now and once more go at it from the top. It does not matter.

I can fail a thousand times, but I only need to do it right once.