I've finished refining my power over the night.
https://users.renyi.hu/~csiszar/Publications/Information_Theory_and_Statistics:_A_Tutorial.pdf
Foundations and Trends in Communications and Information Theory

Let me take a look this for a bit and then I will start programming. Today I might need to spend some time away from the screen, but I'll do what I can in the interim.

It definitely feels like things are starting to move. I'll go from the top and do it properly this time around. It will work without a doubt.

And after it does, by the end of this month I will have made a definite advance towards mastering RL.

For what I want to do, this will enough to get me to the level where it is doable.