"...First, let me modify the AC layer to what I want it to be.
```
inl inner = {action_probs=size; eligibility_decay=1; V=1; scale=1; scale_fin=1; upper=1; mid=1}
```

Yeah, this is it. `scale_fin` is just there to infer the scale on the last episode. For the rest it won't be used. Regardless, it is still a necessary expense.

8:55am.

```
                inl outer = {bias=1; input=sublayer_size}
                inl init =
                    {
                    bias=
                        {
                        action_probs=const zero; eligibility_decay=const zero; V=const zero;
                        scale=const one; scale_fin=const one; upper=const two; mid=const one
                        }
                    input=Struct.map' (inl _ -> const zero) inner
                    }
```

This is good. Actually, let me set the `eligibility_decay` to two. No need to have it directly in the `td` function anymore."