"9:50am.
```
        inl bounded_exp =
            bind <| inl {eta mid upper} ->
                exp (eta * log upper + (val one - eta) * log mid)
```

How do I share custom ops again?

9:55am.

```
inl bounded_exp {eta upper mid} = exp (eta * log upper + (val one - eta) * log mid) // TODO: Make this an Op. Remove the duplicate in Seq.
```

Ugh, this is beyond annoying. I do not want to make the backward pass for this by hand just yet.

Let me keep it as an activation like this and I'll have it pasted in two places for the time being. Later I will make it an Op. But first I want to make sure that I can test it so that if I mess up the backward pass I get some signal of it. Well, I already know what the all the bck for this should be so it won't be a problem.

10am.

```
    inl bounded_exp {eta upper mid} = exp (eta * log upper + (one - eta) * log mid) // TODO: Make this an Op. Remove the duplicate in Seq.
    // Uses KL divergence of univariate Gaussians for cost rather than squared error.
    inl td {r discount_factor eligibility_decay R' V' V scale scale_r scale'} =
        inm eligibility_decay = sigmoid eligibility_decay
        inm {scale scale_r scale'} = sequence_module bounded_exp {scale scale_r scale'}

        inm scale' = scale_r + discount_factor * scale'
        inm R = r + discount_factor * (eligibility_decay * R' + (one - eligibility_decay) * V')
        inm error = R - V
        inm _ = log (scale' / scale) + (sqr scale + sqr error) / sqr scale' - half |> as_cost
        {R scaled_error=primal (error / scale)}
```

Yeah, this is how things should go.

10:05am. I am starting to get engrossed in this.

Now that is done like this I can just set scale' to 0 in the terminal episode.

10:10am. It is beautiful, it all fits together so nicely. I know I am supposed to go and adjust the layer, but I just feel like admiring this for a while."