I am so hopeful about this. This is more than just a stab in the dark. Everything points that this is the ideal formulation of a AC.
...There is something I need to do.

```
out={R'=R; V'=primal V; scale'=scale}
```

That `primal V` is not something that should be done here.

```
inm R = r + discount_factor * (eligibility_decay * R' + (one - eligibility_decay) * primal V')
```

Instead it should be done in the td function. Sure I'll have to pass in one extra argument, but it will be worth it. It is a bad idea to have to care about taking the primal in two different places. Without a doubt it should be here even if I am taking it in a different place...which I won't.

Let me take a break here.