It is finally time for the real. Wouldn't it be great if I could run this and have it work perfectly the same time around?
Ah, let me try it. Most likely I will have to go through some pain, but it will be worth it in the end. At least I am now doing use = a simple linear AC. If it does not work, I will try moving to regular AC and such. It won't be hard to pare down the TD function until it is sensible.