"4:45pm.
```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 265 and 735 out of 1000.
The time was 00:00:02.0525502 for: iteration 0
Starting timing for: iteration 1
Winrate is 315 and 685 out of 1000.
The time was 00:00:02.8598665 for: iteration 1
Starting timing for: iteration 2
Winrate is 419 and 581 out of 1000.
The time was 00:00:03.1405266 for: iteration 2
Starting timing for: iteration 3
Winrate is 493 and 507 out of 1000.
The time was 00:00:02.9388918 for: iteration 3
Starting timing for: iteration 4
Winrate is 522 and 478 out of 1000.
The time was 00:00:03.0129035 for: iteration 4
Starting timing for: iteration 5
Winrate is 547 and 453 out of 1000.
The time was 00:00:02.9910281 for: iteration 5
Starting timing for: iteration 6
Winrate is 569 and 431 out of 1000.
The time was 00:00:03.0078122 for: iteration 6
Starting timing for: iteration 7
Winrate is 580 and 420 out of 1000.
The time was 00:00:02.8172381 for: iteration 7
Starting timing for: iteration 8
Winrate is 591 and 409 out of 1000.
The time was 00:00:02.9480016 for: iteration 8
Starting timing for: iteration 9
Winrate is 564 and 436 out of 1000.
The time was 00:00:03.1854619 for: iteration 9
Starting timing for: iteration 10
Winrate is 614 and 386 out of 1000.
The time was 00:00:02.9017140 for: iteration 10
Starting timing for: iteration 11
Winrate is 603 and 397 out of 1000.
The time was 00:00:02.8937006 for: iteration 11
Starting timing for: iteration 12
Winrate is 606 and 394 out of 1000.
The time was 00:00:02.7546941 for: iteration 12
Starting timing for: iteration 13
Winrate is 625 and 375 out of 1000.
The time was 00:00:02.9272109 for: iteration 13
Starting timing for: iteration 14
Winrate is 622 and 378 out of 1000.
The time was 00:00:02.8846852 for: iteration 14
```

This thing is just so awful. Well, this is the standard AC without KFAC.

4:50pm.

```
The learning_rate is 2 ** -7
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 443 and 557 out of 1000.
The time was 00:00:03.8971915 for: iteration 0
Starting timing for: iteration 1
Winrate is 650 and 350 out of 1000.
The time was 00:00:03.8076309 for: iteration 1
Starting timing for: iteration 2
Winrate is 716 and 284 out of 1000.
The time was 00:00:04.4053032 for: iteration 2
Starting timing for: iteration 3
Winrate is 724 and 276 out of 1000.
The time was 00:00:04.5825516 for: iteration 3
Starting timing for: iteration 4
Winrate is 679 and 321 out of 1000.
The time was 00:00:04.3545934 for: iteration 4
Starting timing for: iteration 5
Winrate is 722 and 278 out of 1000.
The time was 00:00:04.2260006 for: iteration 5
Starting timing for: iteration 6
Winrate is 715 and 285 out of 1000.
The time was 00:00:04.3035364 for: iteration 6
Starting timing for: iteration 7
Winrate is 733 and 267 out of 1000.
The time was 00:00:04.8442116 for: iteration 7
Starting timing for: iteration 8
Winrate is 748 and 252 out of 1000.
The time was 00:00:04.9057153 for: iteration 8
Starting timing for: iteration 9
Winrate is 744 and 256 out of 1000.
The time was 00:00:04.6199654 for: iteration 9
Starting timing for: iteration 10
Winrate is 763 and 237 out of 1000.
The time was 00:00:04.6068284 for: iteration 10
Starting timing for: iteration 11
Winrate is 718 and 282 out of 1000.
The time was 00:00:04.1916014 for: iteration 11
Starting timing for: iteration 12
Winrate is 745 and 255 out of 1000.
The time was 00:00:03.9978827 for: iteration 12
Starting timing for: iteration 13
Winrate is 716 and 284 out of 1000.
The time was 00:00:03.8124127 for: iteration 13
Starting timing for: iteration 14
Winrate is 735 and 265 out of 1000.
The time was 00:00:04.0358197 for: iteration 14
```

Oh, finally. It seems that the way to get this to run it to jack up the learning rates. Now it is firing on both cylinders. Let me see if I can get the KL cost to work this well."