"6:05pm. Suppose I set scale to just `exp(mid * tanh theta)`, what would happen?
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 416 and 584 out of 1000.
The time was 00:00:04.5956751 for: iteration 0
Starting timing for: iteration 1
Winrate is 611 and 389 out of 1000.
The time was 00:00:05.5912613 for: iteration 1
Starting timing for: iteration 2
Winrate is 639 and 361 out of 1000.
The time was 00:00:05.6739867 for: iteration 2
Starting timing for: iteration 3
Winrate is 672 and 328 out of 1000.
The time was 00:00:05.9356857 for: iteration 3
Starting timing for: iteration 4
Winrate is 675 and 325 out of 1000.
The time was 00:00:06.1167339 for: iteration 4
Starting timing for: iteration 5
Winrate is 682 and 318 out of 1000.
The time was 00:00:06.2190723 for: iteration 5
Starting timing for: iteration 6
Winrate is 690 and 310 out of 1000.
The time was 00:00:06.2092471 for: iteration 6
Starting timing for: iteration 7
Winrate is 705 and 295 out of 1000.
The time was 00:00:05.7642333 for: iteration 7
Starting timing for: iteration 8
Winrate is 709 and 291 out of 1000.
The time was 00:00:05.9255021 for: iteration 8
Starting timing for: iteration 9
Winrate is 707 and 293 out of 1000.
The time was 00:00:06.6982561 for: iteration 9
Starting timing for: iteration 10
Winrate is 720 and 280 out of 1000.
The time was 00:00:06.8233662 for: iteration 10
Starting timing for: iteration 11
Winrate is 718 and 282 out of 1000.
The time was 00:00:06.8010936 for: iteration 11
Starting timing for: iteration 12
Winrate is 736 and 264 out of 1000.
The time was 00:00:06.5475094 for: iteration 12
Starting timing for: iteration 13
Winrate is 732 and 268 out of 1000.
The time was 00:00:06.5793002 for: iteration 13
Starting timing for: iteration 14
Winrate is 726 and 274 out of 1000.
The time was 00:00:06.8159463 for: iteration 14
```

Ohh, this is the KL cost now. It seems it like the exponential activation quite a lot.

6:10pm. Let me try dividing the reward by 10 so I can see what happens."