"6:10pm. Let me try dividing the reward by 10 so I can see what happens.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 332 and 668 out of 1000.
The time was 00:00:04.6683617 for: iteration 0
Starting timing for: iteration 1
Winrate is 575 and 425 out of 1000.
The time was 00:00:05.6714794 for: iteration 1
Starting timing for: iteration 2
Winrate is 660 and 340 out of 1000.
The time was 00:00:06.4568369 for: iteration 2
Starting timing for: iteration 3
Winrate is 699 and 301 out of 1000.
The time was 00:00:06.7607356 for: iteration 3
Starting timing for: iteration 4
Winrate is 703 and 297 out of 1000.
The time was 00:00:06.4298835 for: iteration 4
Starting timing for: iteration 5
Winrate is 697 and 303 out of 1000.
The time was 00:00:06.7379978 for: iteration 5
Starting timing for: iteration 6
Winrate is 712 and 288 out of 1000.
The time was 00:00:06.9255156 for: iteration 6
Starting timing for: iteration 7
Winrate is 741 and 259 out of 1000.
The time was 00:00:06.3411765 for: iteration 7
Starting timing for: iteration 8
Winrate is 692 and 308 out of 1000.
The time was 00:00:06.3480231 for: iteration 8
Starting timing for: iteration 9
Winrate is 671 and 329 out of 1000.
The time was 00:00:06.1826066 for: iteration 9
Starting timing for: iteration 10
Winrate is 699 and 301 out of 1000.
The time was 00:00:06.2335951 for: iteration 10
Starting timing for: iteration 11
Winrate is 727 and 273 out of 1000.
The time was 00:00:06.2108046 for: iteration 11
Starting timing for: iteration 12
Winrate is 692 and 308 out of 1000.
The time was 00:00:05.7829098 for: iteration 12
Starting timing for: iteration 13
Winrate is 715 and 285 out of 1000.
The time was 00:00:06.2189822 for: iteration 13
Starting timing for: iteration 14
Winrate is 718 and 282 out of 1000.
The time was 00:00:06.1892475 for: iteration 14
```

Wow it is even more affected by this than the square error. But at least it shows some trend towards stability. Let me set the initial range from 1 to 10 rather than the current 1 to 2.

That is, I'll stop with the `exp tanh`."